+++ RELEASE_NOTES.md
@@ -1,2 +1,3 @@
 v0.67.5
 
+- Fix a bug in the Swift unauth chat listener that caused a crash on connect.